requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
langchain-community>=0.2.0
sentence-transformers>=2.2.2
pdfminer.six>=20221105
python-dotenv>=1.0.0
openai>=0.27.0
faiss-cpu>=1.7.4
